+++ docs/demo/cascader/section-basic.jsx
@@ -3,7 +3,7 @@ import DocViewer from '../../../libs/doc-viewer'
 import Cascader from '../../../components/cascader'
 const prefix = 'section-basic'
 const desc = '展示从多个收起的备选项中选出的一个选项'
-const rightOptions = ['基础', '带默认值', '可清空', '禁用']
+const rightOptions = ['基础', '带默认值', '可清空', '无边框', '禁用']
 const code = [
   {
     code: `import React from 'react'
@@ -157,6 +157,81 @@ class Demo extends React.Component {
   }
 }`,
     opt: ['带默认值']
+  },
+  {
+    code: `import React from 'react'
+import Cascader from '@hi-ui/hiui/es/cascader'\n
+class Demo extends React.Component {
+  constructor () {
+    super()
+    this.state = {
+      options: [
+        {
+          id: '手机',
+          content: '手机',
+          children: [
+            {
+              id: '小米',
+              content: '小米',
+              children: [
+                {
+                  id: '小米3',
+                  content: '小米3'
+                },
+                {
+                  id: '小米4',
+                  content: '小米4'
+                },
+              ]
+            },
+            {
+              id: '红米',
+              content: '红米',
+              children: [
+                {
+                  id: '红米3',
+                  content: '红米3'
+                },
+                {
+                  id: '红米4',
+                  content: '红米4'
+                }
+              ]
+            }
+          ]
+        },
+        {
+          id: '电视',
+          content: '电视',
+          children: [
+            {
+              id: '小米电视4A',
+              content: '小米电视4A'
+            },
+            {
+              id: '小米电视4C',
+              content: '小米电视4C'
+            }
+          ]
+        }
+      ]
+    }
+  }
+  render(){
+    return(
+      <Cascader
+        onChange={(id)=>{
+          console.log('on change', id)
+        }}
+        defaultValue={['电视','小米电视4C']}
+        data={this.state.options}
+        bordered={false}
+        style={{ width: 200 }}
+      />
+    )
+  }
+}`,
+    opt: ['无边框']
   },
   {
     code: `import React from 'react'